split some APIs from nn.py to rnn.py
